requirement.txt
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.49.0
boto3==1.10.34
botocore==1.13.34
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
gensim==3.8.1
html5lib==0.9999999
idna==2.8
jmespath==0.9.4
Markdown==2.2.0
numpy==1.16.2
pandas==0.25.3
protobuf==3.11.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.3
six==1.13.0
smart-open==1.9.0
tensorflow-gpu==1.2.0
toolz==0.10.0
tqdm==4.40.2
urllib3==1.25.7
Werkzeug==0.16.0
